% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wp.R
\name{create_wp}
\alias{create_wp}
\title{Create a new work product}
\usage{
create_wp(wp_dir = "wp", tracker_name = "work-products-tracker.csv")
}
\arguments{
\item{wp_dir}{the name of the work product directory within the project}

\item{tracker_name}{the name of the tracker file in the work product
directory}
}
\description{
Create a new work product in the specified work product directory.
}
\references{
"Guerrilla Analytics: A Practical Approach to Working with Data",
Enda Ridge (\href{https://guerrilla-analytics.net/}{Website})
}
