% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{setup_project}
\alias{setup_project}
\title{Guerrilla Analytics project set up}
\usage{
setup_project(directory, overwrite = FALSE)
}
\arguments{
\item{directory}{The file path to the directory in which you want to create the project.}

\item{overwrite}{If the directory already exists, do you wish to overwrite it and its contents? Defaults to FALSE}
}
\description{
This function creates a basic project skeleton following the Guerilla Analytics structure.
}
\examples{
setup_project("/home/jim/Documents/test_project", overwrite = FALSE)
}

