% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wp.R
\name{create_wp}
\alias{create_wp}
\title{Create a new work product.}
\usage{
create_wp(title, owner, deliver_to, remark, wp_dir = "wp")
}
\arguments{
\item{title}{The title of the work product.}

\item{owner}{The name of the owner of the work product (i.e. the analyst
reponsble).}

\item{deliver_to}{The name of the person the work product will be delivered
to/is for.}

\item{remark}{A brief sentence describing the work product.}

\item{wp_dir}{The name of the work product directory. Defaults to wp.}
}
\description{
This function creates the skeleton for a new work-product within the project.
It must be called from the root directory of the project.
}
\examples{
create_wp("wp", "Customer complaints summary", "Jim Leach", "Enda Ridge", "Summary numbers for customer complaints from the past 6 months, after removing products A, B ande C.")
}
\references{
"Guerrilla Analytics: A Practical Approach to Working with Data", Enda Ridge (\href{https://guerrilla-analytics.net/}{Website})
}

