% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{setup_project}
\alias{setup_project}
\title{Guerrilla Analytics project set up.}
\usage{
setup_project(name, directory, overwrite = FALSE, rstudio_project = FALSE,
  switch_to = FALSE)
}
\arguments{
\item{name}{The name of the project}

\item{directory}{The file path to the directory in which you want to create
the project.}

\item{overwrite}{If the directory already exists, do you wish to overwrite it
and its contents? Defaults to FALSE.}

\item{rstudio_project}{Create an Rstudio project along with the directory?
Defaults to FALSE.}

\item{switch_to}{Switch immediately to the new project? Defaults to FALSE.}
}
\description{
This function creates a basic project skeleton following the Guerilla
Analytics structure.
}
\examples{
setup_project("test_project", "/home/jim/Documents", overwrite = FALSE,
rstudio_project = FALSE, switch_to = FALSE)
setup_project("test_project", "/home/jim/Documents", overwrite = FALSE,
rstudio_project = TRUE, switch_to = FALSE)
}
\references{
"Guerrilla Analytics: A Practical Approach to Working with Data", Enda Ridge (\href{https://guerrilla-analytics.net/}{Website})
}

