% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_wp_tracker.R
\name{view_wp_tracker}
\alias{view_wp_tracker}
\title{View the work products tracker}
\usage{
view_wp_tracker()
}
\description{
A function created purely to facilitate an RStudio addin to display the current work products tracker.
}
\references{
"Guerrilla Analytics: A Practical Approach to Working with Data", Enda Ridge (\href{https://guerrilla-analytics.net/}{Website})
}
